pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-guideline-parser"
version = "0.1.0"
description = "Parse and search medical guidelines from PDF/HTML sources using AI-powered text extraction and BM25 search"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["medical", "guidelines", "pdf-parser", "search", "healthcare", "ai", "bm25", "clinical"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Text Processing :: Markup :: HTML",
    "Topic :: Text Processing :: Markup :: Markdown",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.7,<3",
    "PyMuPDF>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "python-dateutil>=2.9.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    "chardet>=5.2.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=5.0",
    "mypy>=1.0",
]
uvloop = ["uvloop>=0.20.0; platform_system != 'Windows'"]

[project.urls]
Homepage = "https://github.com/yourusername/clinical-guideline-parser"
Documentation = "https://github.com/yourusername/clinical-guideline-parser#readme"
Repository = "https://github.com/yourusername/clinical-guideline-parser.git"
Issues = "https://github.com/yourusername/clinical-guideline-parser/issues"
Changelog = "https://github.com/yourusername/clinical-guideline-parser/blob/main/CHANGELOG.md"

[project.scripts]
clinical-ingest = "src.cli.ingest:main"
clinical-search = "src.cli.search:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.json"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
